Pruning and File Restructuring
