+++ dietfairy/src/components/IngredientCheck/IngredientCheck.js
