+++ dietfairy/src/components/Navbar/Navbar.js
@@ -1,39 +1,56 @@
-import React from "react";
 import React, { Component } from "react";
+import { Navbar, Nav, NavItem } from "react-bootstrap";
+import { LinkContainer } from "react-router-bootstrap";
 
-class Navbar extends Component {
+class TestNavbar extends Component {
   render() {
+    const login_logout = this.props.logged_in ? (
+      <LinkContainer to="/logout">
+        <NavItem eventKey={1}>Logout</NavItem>
+      </LinkContainer>
+    ) : (
+      <LinkContainer to="/login">
+        <NavItem eventKey={1} href="/login">
+          Login
+        </NavItem>
+      </LinkContainer>
+    );
+
+    const profile_register = this.props.logged_in ? (
+      <LinkContainer to="/profile">
+        <NavItem eventKey={2}>Profile</NavItem>
+      </LinkContainer>
+    ) : (
+      <LinkContainer to="/register">
+        <NavItem eventKey={2}>Register</NavItem>
+      </LinkContainer>
+    );
+
+    const upload_noupload = this.props.logged_in ? (
+      <LinkContainer to="/upload">
+        <NavItem eventKey={3}>Upload</NavItem>
+      </LinkContainer>
+    ) : null;
+
     return (
-      <nav className="navbar navbar-expand-lg navbar-light bg-light">
-        <a className="navbar-brand" href="#">
-          Navbar
-        </a>
-        <button
-          className="navbar-toggler"
-          type="button"
-          data-toggle="collapse"
-          data-target="#navbarNavAltMarkup"
-          aria-controls="navbarNavAltMarkup"
-          aria-expanded="false"
-          aria-label="Toggle navigation"
-        >
-          <span className="navbar-toggler-icon"></span>
-        </button>
-        <div className="collapse navbar-collapse" id="navbarNavAltMarkup">
-          <div className="navbar-nav">
-            //home button to be set to home link
-            <a className="nav-item nav-link active" href="#">
-              Home <span className="sr-only">(current)</span>
-            </a>
-            //saved href to be set to final link
-            <a className="nav-item nav-link" href="#">
-              Saved
-            </a>
-          </div>
-        </div>
-      </nav>
+      <div>
+        <Navbar inverse>
+          <Navbar.Header>
+            <Navbar.Brand>
+              <LinkContainer to="/">
+                <a>dietFairy</a>
+              </LinkContainer>
+            </Navbar.Brand>
+          </Navbar.Header>
+          <Nav>
+            {login_logout}
+            {profile_register}
+            {upload_noupload}
+          </Nav>
+        </Navbar>
+      </div>
     );
   }
 }
 
-export default Navbar;
+export default TestNavbar;
