--- dietfairy/src/pages/RecipeSearch.js
@@ -1,1 +1,0 @@
-////recipe search
